Add experimental initial implementation
 * Got fully typed Redux modules (State type, multiple `(type, actionCreator, subReducer)` children) set up
  * Verifies that action types match the key of the object
  * Verifies that return types of action creators match the state
  * Verifies that both a sub-reducer and an action creator are provided
  * Can create a `dispatch` function with arguments made up of the type of all possible actions
  * Verifies that the type of the action returned by the action creator matches the type of the subreducer's `action` argument as well as the key of the object itself

The next step will be to make this functional for multiple levels of modules, just like a normal Redux store.  The created `dispatch` function will have types combined from *all* of the submodules, an aggregate `State` type will be constructed for the entire store, and wrapped versions of `connect` and similar will need to be created.

All in all, I'm incredibly happy that this is working.  The trick of providing a type as a generic parameter and then specializing that type in the argument while also providing `typeof submodule` as the type for the generic type is just the kind of static type-checking that we need.  So hyped.